Refactor higher-functions gcd() and lcm() to use Int.
Refactor Problem 1 & Problem 2.

Update util reusable file docstrings to fit style guideline.

Update docstrings & test method naming convention.

Refactor speed tests to use util higher-functions.

+++ src/main/kotlin/batch0/Problem9.kt
@@ -97,7 +97,7 @@ class SpecialPythagoreanTriplet {
                 }
                 var k = if (m % 2 == 1) m + 2 else m + 1
                 while (k < 2 * m && k <= kMax) {
-                    if (kMax % k == 0 && gcd(k.toLong(), m.toLong()) == 1L) {
+                    if (kMax % k == 0 && gcd(k, m) == 1) {
                         val triplet = pythagoreanTriplet(m, k - m, limit / (k * m))
                         val product = triplet.product()
                         if (product >= maxProduct) {
